# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Verdure Plant Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get single product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist items
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reviews/{productId}:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameFr:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        description:
          type: string
        descriptionFr:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        category:
          $ref: "#/components/schemas/Category"
        featured:
          type: boolean
        careLevel:
          type: string
        lightRequirement:
          type: string
        wateringFrequency:
          type: string
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - nameFr
        - nameAr
        - slug
        - description
        - descriptionFr
        - descriptionAr
        - price
        - stock
        - images
        - featured
        - reviewCount
        - createdAt
    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        nameFr:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionFr:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        featured:
          type: boolean
        careLevel:
          type: string
        lightRequirement:
          type: string
        wateringFrequency:
          type: string
      required:
        - name
        - nameFr
        - nameAr
        - description
        - descriptionFr
        - descriptionAr
        - price
        - stock
    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        nameFr:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionFr:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        featured:
          type: boolean
        careLevel:
          type: string
        lightRequirement:
          type: string
        wateringFrequency:
          type: string
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameFr:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        productCount:
          type: integer
      required:
        - id
        - name
        - nameFr
        - nameAr
        - slug
        - productCount
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        status:
          type: string
        total:
          type: number
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        paymentMethod:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required:
        - id
        - status
        - total
        - customerName
        - customerEmail
        - customerPhone
        - shippingAddress
        - shippingCity
        - paymentMethod
        - items
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - orderId
        - productId
        - productName
        - quantity
        - price
    CreateOrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        paymentMethod:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - customerName
        - customerEmail
        - customerPhone
        - shippingAddress
        - shippingCity
        - paymentMethod
        - items
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
...
    OrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - totalPages
    WishlistItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        createdAt:
          type: string
      required:
        - id
        - userId
        - productId
        - product
        - createdAt
    WishlistInput:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        userName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - productId
        - rating
        - userName
        - createdAt
    CreateReviewInput:
      type: object
      properties:
        rating:
          type: integer
        comment:
          type: string
      required:
        - rating
...