# lib/api-zod/src/generated/types/createReviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Verdure Plant Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewInput {
  rating: number;
  comment?: string;
}