# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Verdure Plant Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface Product {
  id: number;
  name: string;
  nameFr: string;
  nameAr: string;
  slug: string;
  description: string;
  descriptionFr: string;
  descriptionAr: string;
  price: number;
  stock: number;
  images: string[];
  /** @nullable */
  categoryId?: number | null;
  category?: Category;
  featured: boolean;
  careLevel?: string;
  lightRequirement?: string;
  wateringFrequency?: string;
  /** @nullable */
  averageRating?: number | null;
  reviewCount: number;
  createdAt: string;
}